Rename fuzz target
